requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
cohere==4.39
colorama==0.4.6
dnspython==2.4.2
executing==2.0.1
fastavro==1.9.2
frozenlist==1.4.1
icecream==2.1.3
idna==3.6
importlib-metadata==6.11.0
langdetect==1.0.9
multidict==6.0.4
Pygments==2.17.2
pymongo==4.6.1
python-decouple==3.8
requests==2.31.0
six==1.16.0
urllib3==2.1.0
virustotal-api==1.1.11
yarl==1.9.4
zipp==3.17.0
